Remove EccLoaded struct and FixedPoint, FixedPointShort associated types.
We do not load the fixed bases into a single location in the circuit;
rather, they are assigned to fixed columns at the offsets where they
need to be used.

The EccLoaded struct was not storing any references to circuit variables.
It was holding constants in memory which the caller can instead directly
access.

Similarly, there is no need to have the FixedPoint or FixedPointShort
associated types in the EccInstructions trait, since these types do not
contain references to circuit variables. The corresponding get_fixed()
and get_fixed_short() instructions are also removed.