Clarify and widen license exceptions
The intent is to make clear that it is possible to extend the functionality of Zcash implementations (not necessarily only published by the ECC or ZF), and to experiment on testnets, while still relying on the BOSL exception. The leeway given to forks is increased to 3456 blocks (roughly 3 days) to make reasonable allowance for release processes.