update links to orchard design documents
It wasn't clear where within the protocol specification each of
these specific implementation details were defined and stated.
However, the orchard design docs had some great detail as to the
purpose of this, so this was where I decided to link to. In cases
where it was referencing Action commitments the protocol spec
seemed like the best place to link to instead.