Allow SpendValidatingKey to be constructed from bytes
closes #427

This adds a FROST feature flag so that public interface of the
crate is not altered in its meaning for non-FROST applications.

Key derivation is intended to be done on a specific way for usual
applications. Interface changes introduce by FROST can pose a risk
for the rest of the use cases that don't involve the assumptions
of the FROST signature scheme protocol. we don't want non-FROST
cases to make use the helper functions FROST needs to derive the
needed key elements.

PR suggestions:
	Make "frost" feature be under the "unstable" set of features
	correct the unstable feature definition
	use visibility crate